// src/app/models/asset-applying-event.ts
export interface AssetApplyingEvent {
    eventId: string;
    requestOrgIdentifier: string;
    requestOrgNam: string;
    org2: string;
    targetOrgIdentifier: string;
    targetOrgNam: string;
    aggregateId: string;
    status: string;
    message: string;
    assetCategoryId: string;
    assetCategoryThirdLevel: string;
    dateTimeFromNow: string;
}